+++ sdmx-dl-provider-ri/src/main/java/module-info.java
@@ -22,6 +22,7 @@ module sdmxdl.web.ri {
             internal.sdmxdl.ri.web.drivers.FileDriver,
             internal.sdmxdl.ri.web.drivers.InseeDriver2,
             internal.sdmxdl.ri.web.drivers.NbbDriver2,
+            internal.sdmxdl.ri.web.drivers.RngDriver,
             internal.sdmxdl.ri.web.drivers.Sdmx21Driver2;
 
     provides sdmxdl.file.spi.SdmxFileReader with